LMR Bench: 7240857
