+++ src/main.cpp
@@ -25,6 +25,35 @@ namespace nn {
     INCBIN(DefaultNetwork, "corenet.bin");
 }
 
+namespace core {
+    // Declarations
+    Bitboard masks_bit[64], masks_adjacent_file[64], masks_adjacent_north[64], masks_adjacent_south[64], masks_pawn[64][2],
+            masks_knight[64], masks_king[64], masks_file[64], masks_rank[64], masks_rook[64], masks_diagonal[64],
+            masks_anti_diagonal[64], masks_bishop[64], masks_common_ray[64][64];
+    LineType line_type[64][64];
+    Bitboard attack_table_rook[102400], attack_table_bishop[5248];
+} // namespace core
+
+namespace search {
+    Depth lmr_reductions[200][MAX_PLY + 1];
+}
+
+Logger logger;
+
+void init_all() {
+    logger.init("log.txt");
+    logger.info("init_all", "Logger has been initialized");
+
+    core::init_masks();
+    logger.info("init_all", "Bitboard masks have been initialized");
+
+    core::init_magic();
+    logger.info("init_all", "Magic has been initialized");
+
+    search::init_lmr();
+    logger.info("init_all", "LMR has been initialized");
+}
+
 int main(int argc, char *argv[]) {
 
     std::string mode;
